+++ plugins/pcap/main.cpp
@@ -781,13 +781,13 @@ public:
   reader_options(command::opts_builder&& opts) const override {
     return std::move(opts)
       .add<std::string>("interface,i", "network interface to read packets from")
-      .add<size_t>("cutoff,c", "skip flow packets after this many bytes")
-      .add<size_t>("max-flows,m", "number of concurrent flows to track")
-      .add<size_t>("max-flow-age,a", "max flow lifetime before eviction")
-      .add<size_t>("flow-expiry,e", "flow table expiration interval")
-      .add<size_t>("pseudo-realtime-factor,p", "factor c delaying packets by "
+      .add<int64_t>("cutoff,c", "skip flow packets after this many bytes")
+      .add<int64_t>("max-flows,m", "number of concurrent flows to track")
+      .add<int64_t>("max-flow-age,a", "max flow lifetime before eviction")
+      .add<int64_t>("flow-expiry,e", "flow table expiration interval")
+      .add<int64_t>("pseudo-realtime-factor,p", "factor c delaying packets by "
                                                "1/c")
-      .add<size_t>("snaplen", "snapshot length in bytes")
+      .add<int64_t>("snaplen", "snapshot length in bytes")
       .add<double>("drop-rate-threshold", "drop rate that must be exceeded for "
                                           "warnings to occur")
       .add<bool>("disable-community-id", "disable computation of community id "
@@ -817,7 +817,7 @@ public:
   [[nodiscard]] caf::config_option_set
   writer_options(command::opts_builder&& opts) const override {
     return std::move(opts)
-      .add<size_t>("flush-interval,f", "flush to disk after this many packets")
+      .add<int64_t>("flush-interval,f", "flush to disk after this many packets")
       .finish();
   }
